+++ lib/broadcaster/engine.rb
@@ -6,6 +6,7 @@ require 'broadcaster/configuration'
 require 'broadcaster/acts_as_subscriber'
 require 'broadcaster/acts_as_publisher'
 require 'broadcaster/identifiable'
+require 'tote_bag'
 
 ActiveSupport.on_load :active_record do
   ActiveRecord::Base.include Broadcaster::ActsAsPublisher
@@ -16,13 +17,8 @@ module Broadcaster
   class Engine < ::Rails::Engine
     isolate_namespace Broadcaster
 
-    # https://content.pivotal.io/blog/leave-your-migrations-in-your-rails-engines
     initializer :append_migrations do |app|
-      unless app.root.to_s.match root.to_s
-        config.paths['db/migrate'].expanded.each do |expanded_path|
-          app.config.paths['db/migrate'] << expanded_path
-        end
-      end
+      ToteBag::Migration.export(app, config, root)
     end
   end
 end